+++ pkg/api/api/browser/projects/_projectId@string/works/controller.ts
@@ -1,4 +1,4 @@
-import { createWork } from '@violet/api/src/service/browser'
+import { createWork } from '@violet/api/src/service/browser/work'
 import { defineController } from './$relay'
 
 export default defineController(() => ({
